@@ -1,109 +1,110 @@
 {
-  "name": "LWT",
-  "version": "1.0.0",
-  "description": "learning with texts in electron",
-  "main": "app/App.js",
-  "scripts": {
-    "startOld": "tsc --w & electron-forge start",
-    "start": "ng build & concurrently --kill-others \"tsc --w\" \"electron-forge start\"",
-    "package": "electron-forge package",
-    "make": "electron-forge make",
-    "dev": "tsc --w & cd app & node dev"
-  },
-  "repository": "",
-  "keywords": [
-    "Electron",
-    "electron-forge",
-    "nedb",
-    "language-learning"
-  ],
-  "author": "Dano Bozik",
-  "license": "CC0-1.0",
-  "devDependencies": {
-    "@types/nedb": "^1.8.5",
-    "@types/sqlite3": "^3.1.3",
-    "babel-plugin-transform-async-to-generator": "^6.24.1",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.24.1",
-    "concurrently": "^4.0.1",
-    "electron": "^2.0.8",
-    "electron-forge": "^5.2.2",
-    "electron-prebuilt-compile": "2.0.7",
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "~7.3.8",
-    "@angular/compiler-cli": "~7.2.0",
-    "@angular/language-service": "~7.2.0",
-    "@types/node": "~8.9.4",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.0.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.2.2"
-  },
-  "dependencies": {
-    "@angular/animations": "~7.2.0",
-    "@angular/common": "~7.2.0",
-    "@angular/compiler": "~7.2.0",
-    "@angular/core": "~7.2.0",
-    "@angular/forms": "~7.2.0",
-    "@angular/platform-browser": "~7.2.0",
-    "@angular/platform-browser-dynamic": "~7.2.0",
-    "@angular/router": "~7.2.0",
-    "@types/nedb": "^1.8.5",
-    "bootstrap": "4.1.1",
-    "core-js": "^2.5.4",
-    "electron-compile": "^6.4.3",
-    "electron-squirrel-startup": "^1.0.0",
-    "jquery": "^3.3.1",
-    "nedb": "^1.8.0",
-    "ngx-bootstrap": "^5.1.1",
-    "path": "^0.12.7",
-    "reinstall": "^2.0.0",
-    "rxjs": "~6.3.3",
-    "tether": "^1.4.5",
-    "tslib": "^1.9.0",
-    "vue": "^2.5.17",
-    "zone.js": "~0.8.26"
-  },
-  "config": {
-    "forge": {
-      "make_targets": {
-        "win32": [
-          "squirrel"
-        ],
-        "darwin": [
-          "zip"
-        ],
-        "linux": [
-          "deb",
-          "rpm"
-        ]
-      },
-      "electronPackagerConfig": {
-        "packageManager": "npm"
-      },
-      "electronWinstallerConfig": {
-        "name": "LWT"
-      },
-      "electronInstallerDebian": {},
-      "electronInstallerRedhat": {},
-      "github_repository": {
-        "owner": "",
-        "name": ""
-      },
-      "windowsStoreConfig": {
-        "packageName": "",
-        "name": "lwt"
-      }
+    "name": "LWT",
+    "version": "1.0.0",
+    "description": "learning with texts in electron",
+    "main": "app/App.js",
+    "scripts": {
+        "startOld": "tsc --w & electron-forge start",
+        "start": "ng build & concurrently --kill-others \"tsc --w\" \"electron-forge start\"",
+        "package": "electron-forge package",
+        "make": "electron-forge make",
+        "dev": "tsc --w & cd app & node dev",
+        "test": "jest --verbose",
+        "electron": "concurrently --kill-others \"tsc --w\" \"electron-forge start --inspect-electron\""
+    },
+    "repository": "",
+    "keywords": [
+        "Electron",
+        "electron-forge",
+        "nedb",
+        "language-learning"
+    ],
+    "author": "Dano Bozik",
+    "license": "CC0-1.0",
+    "devDependencies": {
+        "@angular-devkit/build-angular": "~0.13.0",
+        "@angular/cli": "~7.3.8",
+        "@angular/compiler-cli": "~7.2.0",
+        "@angular/language-service": "~7.2.0",
+        "@types/jest": "^24.0.18",
+        "@types/nedb": "^1.8.5",
+        "@types/node": "^12.7.11",
+        "babel-plugin-transform-async-to-generator": "^6.24.1",
+        "babel-preset-env": "^1.7.0",
+        "babel-preset-react": "^6.24.1",
+        "codelyzer": "~4.5.0",
+        "concurrently": "^4.0.1",
+        "electron": "^6.0.11",
+        "electron-forge": "^5.2.2",
+        "electron-prebuilt-compile": "2.0.7",
+        "jasmine-core": "~2.99.1",
+        "jasmine-spec-reporter": "~4.2.1",
+        "jest": "^24.9.0",
+        "karma": "~4.0.0",
+        "karma-chrome-launcher": "~2.2.0",
+        "karma-coverage-istanbul-reporter": "~2.0.1",
+        "karma-jasmine": "~1.1.2",
+        "karma-jasmine-html-reporter": "^0.2.2",
+        "protractor": "~5.4.0",
+        "ts-jest": "^24.1.0",
+        "ts-node": "~7.0.0",
+        "tslint": "~5.11.0",
+        "typescript": "^3.2.2"
+    },
+    "dependencies": {
+        "@angular/animations": "~7.2.0",
+        "@angular/common": "~7.2.0",
+        "@angular/compiler": "~7.2.0",
+        "@angular/core": "~7.2.0",
+        "@angular/forms": "~7.2.0",
+        "@angular/platform-browser": "~7.2.0",
+        "@angular/platform-browser-dynamic": "~7.2.0",
+        "@angular/router": "~7.2.0",
+        "bootstrap": "^4.3.1",
+        "core-js": "^2.5.4",
+        "electron-compile": "^6.4.3",
+        "electron-squirrel-startup": "^1.0.0",
+        "jquery": "^3.3.1",
+        "nedb": "^1.8.0",
+        "ngx-bootstrap": "^5.1.1",
+        "path": "^0.12.7",
+        "reinstall": "^2.0.0",
+        "rxjs": "~6.3.3",
+        "tether": "^1.4.5",
+        "tslib": "^1.9.0",
+        "vue": "^2.5.17",
+        "zone.js": "~0.8.26"
+    },
+    "config": {
+        "forge": {
+            "make_targets": {
+                "win32": [
+                    "squirrel"
+                ],
+                "darwin": [
+                    "zip"
+                ],
+                "linux": [
+                    "deb",
+                    "rpm"
+                ]
+            },
+            "electronPackagerConfig": {
+                "packageManager": "npm"
+            },
+            "electronWinstallerConfig": {
+                "name": "LWT"
+            },
+            "electronInstallerDebian": {},
+            "electronInstallerRedhat": {},
+            "github_repository": {
+                "owner": "",
+                "name": ""
+            },
+            "windowsStoreConfig": {
+                "packageName": "",
+                "name": "lwt"
+            }
+        }
     }
-  }
 }
