@@ -1,4 +1,4 @@
-import * as Objects from './namespace';
+import * as Objects from '.';
 
 export class Text {
     // text
@@ -9,11 +9,17 @@ export class Text {
     public title: string;
     // language
     public languageId: string;
+
+    public userId?: string;
     // word list
-    public words: Objects.WordObject[];
+    public words?: Objects.WordObject[];
     // parsed text
     // sentences list
-    public sentences: string[];
+    public sentences?: string[];
+
+    public _id?: string;
+    public createdOn?: Date;
+    public textParts?: Objects.TextPart[];
 
     constructor() {
     }
