@@ -8,8 +8,8 @@ function addText(): void {
     const text = $('#text').val();
     const title = $('#title').val();
 
-    const textService = new Services.TextService();
-    textService.saveText(text, title, '1', '1').subscribe(savedText => {
-        ipcRenderer.send('main-open-text', savedText._id);
-    });
+    // const textService = new Services.TextService();
+    // textService.saveText(text, title, '1', '1').subscribe(savedText => {
+    //     ipcRenderer.send('main-open-text', savedText._id);
+    // });
 }
