@@ -9,6 +9,9 @@ import { AppComponent } from './app.component';
 import { LoginComponent } from './login/login.component';
 import { SettingsComponent } from './settings/settings.component';
 import { SignupComponent } from './signup/signup.component';
+import { TextsComponent } from './texts/texts.component';
+import { ReadTextComponent } from './read-text/read-text.component';
+import { WebviewDirective } from './webview.directive';
 
 @NgModule({
     declarations: [
@@ -16,7 +19,10 @@ import { SignupComponent } from './signup/signup.component';
         LoginComponent,
         AddTextComponent,
         SignupComponent,
-        SettingsComponent
+        SettingsComponent,
+        TextsComponent,
+        ReadTextComponent,
+        WebviewDirective,
     ],
     imports: [
         BrowserModule,
